New Check Added : test_apigateway_rest_api_waf_acl_attached
